pyproject.toml
[project]
name = "valoafkpickbot"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "mao", email = "90289410+injectxr@users.noreply.github.com" }
]
dependencies = [
    "discord-py>=2.4.0",
    "valclient>=2.28",
    "requests>=2.32.3",
    "python-dotenv>=1.0.1",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/valoafkpickbot"]
